Backend Rework for Auth

+++ perpustakaan-backend/app/Services/Denda/DendaService.php
@@ -4,55 +4,53 @@ namespace App\Services\Denda;
 
 use App\Models\Denda;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 
-class DendaService
+class DendaService implements HasMiddleware
 {
+    public static function middleware()
+    {
+        return [
+            new Middleware('auth:sanctum', except: ['index', 'show'])
+        ];
+    }
+
     public function create(Request $request)
     {
-        $request->validate([
+        $fields = $request->validate([
             'hari' => 'required|integer',
             'harga_denda' => 'required|integer',
             'status_pembayaran' => 'required|in:sudah dibayar,belum dibayar',
         ]);
 
-        return Denda::create($request->all());
+        $denda = $request->user()->denda()->create($fields);
+
+        return $denda;
     }
 
-    public function getAll()
+    public function update(Request $request, Denda $denda)
     {
-        return Denda::all();
-    }
+        Gate::authorize('modify', $denda);
 
-    public function getTotalDenda()
-    {
-        $totalDenda = Denda::count();
-        return response()->json(['total' => $totalDenda]);
-    }
-
-    public function getById($id)
-    {
-        return Denda::findOrFail($id);
-    }
-
-    public function update($id, Request $request)
-    {
-        $request->validate([
+        $fields = $request->validate([
             'hari' => 'sometimes|integer',
             'harga_denda' => 'sometimes|integer',
             'status_pembayaran' => 'sometimes|in:sudah dibayar,belum dibayar',
         ]);
 
-        $denda = Denda::findOrFail($id);
-        $denda->update($request->all());
+        $denda->update($fields);
 
         return $denda;
     }
 
-    public function delete($id)
+    public function delete(Denda $denda)
     {
-        $denda = Denda::findOrFail($id);
+        Gate::authorize('modify', $denda);
+
         $denda->delete();
 
-        return response()->json(['message' => 'Denda berhasil dihapus']);
+        return ['message' => 'Denda berhasil dihapus'];
     }
 }
